Add id as cache id
