+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './connectionUpdater';
 export * from './getNodes';
 export * from './relayStylePagination';
 export * from './mutationUpdater';
+export * from './idAsCacheId';
